+++ recfm-java-addon/src/main/java/io/github/epi155/recfm/type/FieldAbc.java
@@ -6,10 +6,12 @@ import io.github.epi155.recfm.api.CheckAbc;
 import io.github.epi155.recfm.api.NormalizeAbcMode;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
+import lombok.extern.slf4j.Slf4j;
 import lombok.val;
 
 @Data
 @EqualsAndHashCode(callSuper = true)
+@Slf4j
 public class FieldAbc extends FloatingField implements AbcModel {
     private char padChar = ' ';
     private CheckAbc check;
@@ -41,6 +43,7 @@ public class FieldAbc extends FloatingField implements AbcModel {
         res.setOverride(isOverride());
         res.setLength(getLength());
         res.setOffset(getOffset() + plus);
+        log.debug("  -- {} @{}+{} -> {}", getName(), getOffset(), getLength(), res.getOffset());
         return res;
     }
 }