+++ recfm-java-addon/src/main/java/io/github/epi155/recfm/type/TraitDefine.java
@@ -14,17 +14,25 @@ import java.io.File;
 import java.io.IOException;
 import java.io.PrintWriter;
 import java.util.ArrayList;
+import java.util.LinkedList;
 import java.util.List;
 
 @Data
 @Slf4j
 public class TraitDefine implements ParentFields, TraitModel {
+    private static final List<String> LIST = new LinkedList<>();
     private String name;
     private int length;
     private List<FieldModel> fields = new ArrayList<>();
     private Boolean doc;
+    private final List<String> implementsList = new LinkedList<>();
 
     protected static final String DOT_JAVA = ".java";
+
+    public static boolean contains(String name) {
+        return LIST.contains(name);
+    }
+
     @Override
     public void create(String namespace, GenerateArgs ga, FieldDefault defaults) {
         if (getFields().isEmpty()) return;
@@ -53,11 +61,17 @@ public class TraitDefine implements ParentFields, TraitModel {
                 throw new ClassDefineException(e);
             }
             log.info("  [######] Created.");
+            LIST.add(getName());
         } else {
             throw new ClassDefineException("Class <" + getName() + "> bad defined");
         }
 
 
 
+    }
+
+    @Override
+    public void addImplements(String interfaceName) {
+        implementsList.add(interfaceName);
     }
 }